refactor: relocate repository types out of legacy stores
